style-correction
organised code, corrected styles, fixed some bugs
