+++ css/style.css
@@ -1,7 +1,6 @@
 @import 'https://unpkg.com/swiper@8/swiper-bundle.min.css';
 @import './font.css';
 @import './utilities.css';
-@import './responsive.css';
 
 *,
 *::before,
@@ -149,7 +148,7 @@ body{
 /* term cards styles */
 .term-cards-wrapper{
     display: flex;
-    justify-content: space-around;
+    justify-content: flex-start;
     align-items: center;
     flex-wrap: wrap;
     width: 100%;
@@ -511,7 +510,11 @@ body{
 }
 .container .about-block h2 {
     display: inline-block;
-    font: normal normal 400 16px/19px BPG-ghalo-caps-regular;
+    /* font: normal normal 400 16px/19px BPG-ghalo-caps-regular; */
+    font-size: var(--large);
+    font-weight: bold;
+    color: var(--dark-text);
+    line-height: 1.5;
     margin-bottom: 10px;
     border-bottom: 3px solid #8A8FA3;
 }
@@ -603,39 +606,73 @@ body{
 .alphabet-wrapper .al-active {
     color: var(--card-bg);
 }
+
+/* SELECT AND DROPDOWN STYLES */
 .container .search-block .searchbar form select {
     font-size: var(--medium);
     line-height: 19px;
     height: 25px;
     padding-left: 10px;
     outline: none;
-    border-radius: 0%;
-    border: 1px var(--main-font) solid;
-    color: var(--main-font);
-}
-#filter {
-    width: 250px;
-    position: absolute;
-    left: 73%;
-    top: 50%;
-    transform: translate(-73%, -50%);
-}
-select{
+    font-family: 'firaGo-regular';
+    font-weight: 600;
     cursor: pointer;
+    appearance: none;
 }
-#sort {
-    width: 144px;
+
+.select-wrapper {
+    /* display: grid; */
+    /* grid-template-columns: .75fr .5fr; */
+    display: flex;
+    justify-content: flex-end;
+    width: 100%;
     position: absolute;
-    left: 95%;
     top: 50%;
-    transform: translate(-95%, -50%);
+    right: 1rem;
+    transform: translateY(-50%);
+    width: 40%;
 }
+
+.select {
+    position: relative;
+    display: inline-block;
+}
+
+.custom-arrow {
+    position: absolute;
+    top: 50%;
+    right: 1.4rem;
+    transform: translateY(-50%);
+    pointer-events: none;
+}
+
+.custom-arrow.last {
+    right: .2rem;
+}
+
 #filter, #sort {
+    /* width: 100%; */
+    padding-right: 1.2rem;
     border-radius: 5px;
     background-color: #F5F8FD;
     border: none;
     color: #343645;
 }
+
+#filter {
+    width: 240px;
+    margin-right: 1.25rem;
+}
+#sort {
+    width: 144px;
+}
+
+option {
+    color: var(--main-font);
+    font-family: 'firaGo-regular';
+    font-weight: bold;
+}
+
 .container .search-block input[type=text] {
     width: 100%;
     height: 100%;
@@ -646,21 +683,33 @@ select{
     border-radius: 10px;
     border: 1.5px solid var(--details);
 }
+
 /* slick slider on about page */
-.slick-prev:before {
-    content: "<";
-    color: var(--main-font);
-    font-size: 30px;
+.slick-prev,
+.slick-next {
+    border-radius: 50%;
+    background-color: rgba(0, 0, 0, 0.11);
+}
+.slick-prev:before,
+.slick-next:before {
+    content: url'(../media/png/btn-left-large.png)';
+    color: rgb(177, 177, 177);
+    width: 100%;
+    height: 100%;
+    font-size: 24px;
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
 }
 .slick-next:before {
-    content: ">";
-    color: var(--main-font);
-    font-size: 30px;
+    content: url'(../media/png/btn-right-large.png)';
 }
 .slick-slide-wrapper {
     position: relative;
     height: 260px;
     margin-right: 16px;
+    margin-bottom: 30px;
     background-size: cover;
     background-repeat: no-repeat;
     background-position: center;
@@ -673,7 +722,7 @@ select{
     height: 100%;
     width: 100%;
     opacity: 0;
-    background: rgba(0,0,0,.5);
+    background: rgba(0,0,0,.3);
     -moz-transition: all .3s;
     -webkit-transition: all .3s;
     transition: all .3s;
@@ -683,6 +732,22 @@ select{
 }
 .slick-prev {
     z-index: 22222;
+}
+.slick-slide-wrapper:hover .about-person {
+    opacity: 1;
+}
+.slick-slide-wrapper:hover::after {
+    opacity: 1;
+}
+
+.slick-dots li button {
+    border-radius: 50%;
+    background-color: var(--details);
+    width: 16px;
+    height: 16px;
+}
+.slick-dots li button::before {
+    display: none;
 }
 .about-person {
     display: flex;
@@ -700,12 +765,6 @@ select{
     color: #fff;
     transition: .3s ease-out;
     opacity: 0;
-}
-.slick-slide-wrapper:hover .about-person {
-    opacity: 1;
-}
-.slick-slide-wrapper:hover::after {
-    opacity: 1;
 }
 .about-person span:not(:last-child) {
     margin-bottom: 10px;
@@ -774,8 +833,9 @@ select{
     width: 80%;
 }
 /* swiper */
-.swiper-wrapper{
-    width: 83%;
+.swiper-container {
+    position: relative;
+    /* height: 20px; */
 }
 .swiper {
     margin-top: 10px;
@@ -784,28 +844,44 @@ select{
     display: none;
     padding-left: 35px;
     padding-right: 35px;
-  }
+}
+.swiper-wrapper{
+    /* width: 100%; */
+    width: 83%;
+}
 .swiper-slide {
     cursor: pointer;
     text-align: center;
     width: calc((100%)/10) !important;
+    /* background-color: red; */
 }
 .swiper-horizontal .swiper-scrollbar {
     bottom: 1px;
 }
 .swiper-button-next {
-    right: 0;
+    right: -1.5rem;
     color: var(--card-bg);
+    background: url(../media/png/btn-right.png);
 }
 .swiper-button-prev {
-    left: 0;
+    left: -1.5rem;
     color: var(--card-bg);
+    background: url(../media/png/btn-left.png);
 }
+
 .swiper-button-next::after,
 .swiper-button-prev::after {
-    line-height: 20px;
-    overflow: hidden;
+    display: none;
 }
+
+.swiper-button-prev,
+.swiper-button-next {
+    display: none;
+    background-position: center;
+    background-size: contain;
+    background-size: cover;
+}
+
 .swiper-wrapper div.active-letter{
     font-weight: bold;
     color: #3F4C5F;
@@ -816,4 +892,3 @@ select{
 .swiper-scrollbar-drag {
     background: #D4E2ED;
 }
-/* swiper */
