+++ html/about.html
@@ -18,7 +18,7 @@
     <!-- <link rel="stylesheet" href="../css/font.css"> -->
     <!-- <link rel="stylesheet" href="../css/utilities.css"> -->
     <link rel="stylesheet" href="../css/style.css">
-    <!-- <link rel="stylesheet" href="../css/responsive.css"> -->
+    <link rel="stylesheet" href="../css/responsive.css">
     <link rel="shortcut icon" href="favicon.ico">
     <title>უნილაბის ლექსიკონი</title>
 </head>
@@ -47,7 +47,7 @@
     </header>
     <main class="container">
         <article class="about-block">
-            <h2>პროექტი შესახებ</h2>
+            <h2>პროექტის შესახებ</h2>
             <p>
                 Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
             </p>
@@ -65,10 +65,8 @@
             </div>
         </div>
     </footer>
-    <!-- <script type="text/javascript" src="//code.jquery.com/jquery-1.11.0.min.js"></script> -->
-    <!-- <script type="text/javascript" src="//code.jquery.com/jquery-migrate-1.2.1.min.js"></script> -->
-    <!-- <script type="text/javascript" src="../js/slick/slick.min.js"></script> -->
-    <script type="module" src="../js/main.js"></script>
+    <script type="text/javascript" src="//code.jquery.com/jquery-1.11.0.min.js"></script>
+    <script type="module" src="../js/about.js"></script>
 </body>
 
 </html>
