+++ html/error-page.html
@@ -15,7 +15,7 @@
     <!-- <link rel="stylesheet" href="../css/font.css"> -->
     <!-- <link rel="stylesheet" href="../css/utilities.css"> -->
     <link rel="stylesheet" href="../css/style.css">
-    <!-- <link rel="stylesheet" href="../css/responsive.css"> -->
+    <link rel="stylesheet" href="../css/responsive.css">
     <link rel="shortcut icon" href="favicon.ico">
     <title>უნილაბის ლექსიკონი</title>
 </head>
