+++ index.html
@@ -16,7 +16,7 @@
     <!-- <link rel="stylesheet" href="css/font.css"> -->
     <!-- <link rel="stylesheet" href="css/utilities.css"> -->
     <link rel="stylesheet" href="css/style.css">
-    <!-- <link rel="stylesheet" href="css/responsive.css"> -->
+    <link rel="stylesheet" href="css/responsive.css">
     <link rel="shortcut icon" href="favicon.ico">
     <title>უნილაბის ლექსიკონი</title>
 </head>
@@ -57,8 +57,8 @@
             </div>
         </article>
         <div class="main-input-wrapper">
-            <form class="input-form" action="">
-                <input class="main-input" type="text" placeholder="ძიება">
+            <form action="./html/dictionary.html" class="input-form" action="">
+                <input class="main-input" type="text" placeholder="ძიება" id="main-search" required>
             </form>
             <svg class="search-icon" width="24" height="24" xmlns="http://www.w3.org/2000/svg" fill-rule="evenodd"
                 clip-rule="evenodd">
