+++ js/data.js
@@ -1,0 +1,325 @@
+export const members = [
+  {
+    memberId: 0,
+    memberName: "სახელი",
+    memberLastname: "გვარი",
+    memberPosition: "Frontend Developer",
+    memberImagePath: "../media/png/person.png",
+  },
+  {
+    memberId: 1,
+    memberName: "სახელი",
+    memberLastname: "გვარი",
+    memberPosition: "Backend Developer",
+    memberImagePath: "../media/png/person.png",
+  },
+  {
+    memberId: 2,
+    memberName: "სახელი",
+    memberLastname: "გვარი",
+    memberPosition: "UI/UX Designer",
+    memberImagePath: "../media/png/person.png",
+  },
+  {
+    memberId: 3,
+    memberName: "სახელი",
+    memberLastname: "გვარი",
+    memberPosition: "Project Manager",
+    memberImagePath: "../media/png/person.png",
+  },
+];
+
+export const data = [
+  {
+    id: 0,
+    titleEng: "Project management",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description:
+      "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "frontend"],
+    iconPath: "../media/svg/term-cards/term-icon-1.svg",
+    keyword: "frontend",
+  },
+  {
+    id: 1,
+    titleEng: "Project manager",
+    titleGeo: "პროექტის მენეჯერი",
+    Description: " PM, ერთ-ერთი მთავარი მონაწილე პროექტში. მისი როლი არის",
+    hashTags: ["გრაფიკულიდიზაინი", "wordpress"],
+    iconPath: "../media/svg/term-cards/term-icon-2.svg",
+    keyword: "wordpress",
+  },
+  {
+    id: 2,
+    titleEng: "ICT Project",
+    titleGeo: "ICT პროექტი",
+    Description:
+      "ICT იშიფრება, როგორც ინფორმაციული და საკომუნიკაციო როგორც ინფორმაციული და საკომუნიკაციო",
+    hashTags: ["გრაფიკულიდიზაინი", "python"],
+    iconPath: "../media/svg/term-cards/term-icon-3.svg",
+    keyword: "python",
+  },
+  {
+    id: 3,
+    titleEng: "Project management 3",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "frontend"],
+    iconPath: "../media/svg/term-cards/term-icon-1.svg",
+    keyword: "frontend",
+  },
+  {
+    id: 4,
+    titleEng: "Project management 4",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "wordpress"],
+    iconPath: "../media/svg/term-cards/term-icon-2.svg",
+    keyword: "wordpress",
+  },
+  {
+    id: 5,
+    titleEng: "Project management 5",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "python"],
+    iconPath: "../media/svg/term-cards/term-icon-3.svg",
+    keyword: "python",
+  },
+  {
+    id: 6,
+    titleEng: "Project management 6",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "frontend"],
+    iconPath: "../media/svg/term-cards/term-icon-1.svg",
+    keyword: "frontend",
+  },
+  {
+    id: 7,
+    titleEng: "Project management 7",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "wordpress"],
+    iconPath: "../media/svg/term-cards/term-icon-2.svg",
+    keyword: "wordpress",
+  },
+  {
+    id: 8,
+    titleEng: "Project management 8",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "python"],
+    iconPath: "../media/svg/term-cards/term-icon-3.svg",
+    keyword: "python",
+  },
+  {
+    id: 9,
+    titleEng: "Project management 9",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "python"],
+    iconPath: "../media/svg/term-cards/term-icon-1.svg",
+    keyword: "python",
+  },
+  {
+    id: 10,
+    titleEng: "Project management 10",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "wordpress"],
+    iconPath: "../media/svg/term-cards/term-icon-2.svg",
+    keyword: "wordpress",
+  },
+  {
+    id: 11,
+    titleEng: "Project management 11",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "frontend"],
+    iconPath: "../media/svg/term-cards/term-icon-3.svg",
+    keyword: "frontend",
+  },
+  {
+    id: 12,
+    titleEng: "Project management 12",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "python"],
+    iconPath: "../media/svg/term-cards/term-icon-1.svg",
+    keyword: "python",
+  },
+  {
+    id: 13,
+    titleEng: "Project management 13",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "wordpress"],
+    iconPath: "../media/svg/term-cards/term-icon-2.svg",
+    keyword: "wordpress",
+  },
+  {
+    id: 14,
+    titleEng: "Project management 14",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "frontend"],
+    iconPath: "../media/svg/term-cards/term-icon-3.svg",
+    keyword: "frontend",
+  },
+  {
+    id: 15,
+    titleEng: "Project management 15",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "python"],
+    iconPath: "../media/svg/term-cards/term-icon-1.svg",
+    keyword: "python",
+  },
+  {
+    id: 16,
+    titleEng: "Project management 16",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "wordpress"],
+    iconPath: "../media/svg/term-cards/term-icon-2.svg",
+    keyword: "wordpress",
+  },
+  {
+    id: 17,
+    titleEng: "Project management 17",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "frontend"],
+    iconPath: "../media/svg/term-cards/term-icon-3.svg",
+    keyword: "frontend",
+  },
+  {
+    id: 18,
+    titleEng: "Project management 18",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "python"],
+    iconPath: "../media/svg/term-cards/term-icon-1.svg",
+    keyword: "python",
+  },
+  {
+    id: 19,
+    titleEng: "Project management 19",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "wordpress"],
+    iconPath: "../media/svg/term-cards/term-icon-2.svg",
+    keyword: "wordpress",
+  },
+  {
+    id: 20,
+    titleEng: "Project management 20",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "frontend"],
+    iconPath: "../media/svg/term-cards/term-icon-3.svg",
+    keyword: "frontend",
+  },
+  {
+    id: 21,
+    titleEng: "Project management 21",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "python"],
+    iconPath: "../media/svg/term-cards/term-icon-1.svg",
+    keyword: "python",
+  },
+  {
+    id: 22,
+    titleEng: "Project management 22",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "wordpress"],
+    iconPath: "../media/svg/term-cards/term-icon-2.svg",
+    keyword: "wordpress",
+  },
+  {
+    id: 23,
+    titleEng: "Project management 23",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "frontend"],
+    iconPath: "../media/svg/term-cards/term-icon-3.svg",
+    keyword: "frontend",
+  },
+  {
+    id: 24,
+    titleEng: "Project management 24",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "python"],
+    iconPath: "../media/svg/term-cards/term-icon-1.svg",
+    keyword: "python",
+  },
+  {
+    id: 25,
+    titleEng: "Project management 25",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "wordpress"],
+    iconPath: "../media/svg/term-cards/term-icon-2.svg",
+    keyword: "wordpress",
+  },
+  {
+    id: 26,
+    titleEng: "Project management 26",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "frontend"],
+    iconPath: "../media/svg/term-cards/term-icon-3.svg",
+    keyword: "frontend",
+  },
+  {
+    id: 27,
+    titleEng: "Project management 27",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "python"],
+    iconPath: "../media/svg/term-cards/term-icon-1.svg",
+    keyword: "python",
+  },
+  {
+    id: 28,
+    titleEng: "Project management 28",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "wordpress"],
+    iconPath: "../media/svg/term-cards/term-icon-2.svg",
+    keyword: "wordpress",
+  },
+  {
+    id: 29,
+    titleEng: "Project management 29",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "frontend"],
+    iconPath: "../media/svg/term-cards/term-icon-3.svg",
+    keyword: "frontend",
+  },
+  {
+    id: 30,
+    titleEng: "Project management 30",
+    titleGeo: "პროექტის მენეჯმენტი",
+    Description: "ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა",
+    hashTags: ["გრაფიკულიდიზაინი", "frontend"],
+    iconPath: "../media/svg/term-cards/term-icon-3.svg",
+    keyword: "frontend",
+  },
+  {
+    id: 31,
+    titleEng: "Project",
+    titleGeo: "პროექტი",
+    Description:
+      "პროექტი არის დროებითი საქმიანობა უნიკალური პროდუქტის, მომსახურების ან შედეგის შექმნის მიზნით",
+    hashTags: ["ციფრულიკომუნიკაციები", "პროექტისმენეჯმენტი"],
+    iconPath: "../media/svg/term-cards/term-icon-3.svg",
+    keyword: "frontend",
+    weekTerm: true,
+  },
+];
