+++ js/main.js
@@ -1,706 +1,113 @@
-const data = [
-    {
-        id: 0,
-        titleEng: 'Project management',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'frontend'],
-        iconPath: 'media/svg/term-cards/term-icon-1.svg',
-        keyword: 'frontend'
-    },
-    {
-        id: 1,
-        titleEng: 'Project manager',
-        titleGeo: 'პროექტის მენეჯერი',
-        Description: ' PM, ერთ-ერთი მთავარი მონაწილე პროექტში. მისი როლი არის',
-        hashTags: ['გრაფიკულიდიზაინი', 'wordpress'],
-        iconPath: 'media/svg/term-cards/term-icon-2.svg',
-        keyword: 'wordpress'
-    },
-    {
-        id: 2,
-        titleEng: 'ICT Project',
-        titleGeo: 'ICT პროექტი',
-        Description: 'ICT იშიფრება, როგორც ინფორმაციული და საკომუნიკაციო როგორც ინფორმაციული და საკომუნიკაციო',
-        hashTags: ['გრაფიკულიდიზაინი', 'python'],
-        iconPath: 'media/svg/term-cards/term-icon-3.svg',
-        keyword: 'python'
-    },
-    {
-        id: 3,
-        titleEng: 'Project management 3',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'frontend'],
-        iconPath: 'media/svg/term-cards/term-icon-1.svg',
-        keyword: 'frontend'
-    },
-    {
-        id: 4,
-        titleEng: 'Project management 4',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'wordpress'],
-        iconPath: 'media/svg/term-cards/term-icon-2.svg',
-        keyword: 'wordpress'
-    },
-    {
-        id: 5,
-        titleEng: 'Project management 5',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'python'],
-        iconPath: 'media/svg/term-cards/term-icon-3.svg',
-        keyword: 'python'
-    },
-    {
-        id: 6,
-        titleEng: 'Project management 6',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'frontend'],
-        iconPath: 'media/svg/term-cards/term-icon-1.svg',
-        keyword: 'frontend'
-    },
-    {
-        id: 7,
-        titleEng: 'Project management 7',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'wordpress'],
-        iconPath: 'media/svg/term-cards/term-icon-2.svg',
-        keyword: 'wordpress'
-    },
-    {
-        id: 8,
-        titleEng: 'Project management 8',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'python'],
-        iconPath: 'media/svg/term-cards/term-icon-3.svg',
-        keyword: 'python'
-    },
-    {
-        id: 9,
-        titleEng: 'Project management 9',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'python'],
-        iconPath: 'media/svg/term-cards/term-icon-1.svg',
-        keyword: 'python'
-    },
-    {
-        id: 10,
-        titleEng: 'Project management 10',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'wordpress'],
-        iconPath: 'media/svg/term-cards/term-icon-2.svg',
-        keyword: 'wordpress'
-    },
-    {
-        id: 11,
-        titleEng: 'Project management 11',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'frontend'],
-        iconPath: 'media/svg/term-cards/term-icon-3.svg',
-        keyword: 'frontend'
-    },
-    {
-        id: 12,
-        titleEng: 'Project management 12',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'python'],
-        iconPath: 'media/svg/term-cards/term-icon-1.svg',
-        keyword: 'python'
-    },
-    {
-        id: 13,
-        titleEng: 'Project management 13',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'wordpress'],
-        iconPath: 'media/svg/term-cards/term-icon-2.svg',
-        keyword: 'wordpress'
-    },
-    {
-        id: 14,
-        titleEng: 'Project management 14',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'frontend'],
-        iconPath: 'media/svg/term-cards/term-icon-3.svg',
-        keyword: 'frontend'
-    },
-    {
-        id: 15,
-        titleEng: 'Project management 15',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'python'],
-        iconPath: 'media/svg/term-cards/term-icon-1.svg',
-        keyword: 'python'
-    },
-    {
-        id: 16,
-        titleEng: 'Project management 16',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'wordpress'],
-        iconPath: 'media/svg/term-cards/term-icon-2.svg',
-        keyword: 'wordpress'
-    },
-    {
-        id: 17,
-        titleEng: 'Project management 17',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'frontend'],
-        iconPath: 'media/svg/term-cards/term-icon-3.svg',
-        keyword: 'frontend'
-    },
-    {
-        id: 18,
-        titleEng: 'Project management 18',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'python'],
-        iconPath: 'media/svg/term-cards/term-icon-1.svg',
-        keyword: 'python'
-    },
-    {
-        id: 19,
-        titleEng: 'Project management 19',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'wordpress'],
-        iconPath: 'media/svg/term-cards/term-icon-2.svg',
-        keyword: 'wordpress'
-    },
-    {
-        id: 20,
-        titleEng: 'Project management 20',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'frontend'],
-        iconPath: 'media/svg/term-cards/term-icon-3.svg',
-        keyword: 'frontend'
-    },
-    {
-        id: 21,
-        titleEng: 'Project management 21',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'python'],
-        iconPath: 'media/svg/term-cards/term-icon-1.svg',
-        keyword: 'python'
-    },
-    {
-        id: 22,
-        titleEng: 'Project management 22',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'wordpress'],
-        iconPath: 'media/svg/term-cards/term-icon-2.svg',
-        keyword: 'wordpress'
-    },
-    {
-        id: 23,
-        titleEng: 'Project management 23',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'frontend'],
-        iconPath: 'media/svg/term-cards/term-icon-3.svg',
-        keyword: 'frontend'
-    },
-    {
-        id: 24,
-        titleEng: 'Project management 24',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'python'],
-        iconPath: 'media/svg/term-cards/term-icon-1.svg',
-        keyword: 'python'
-    },
-    {
-        id: 25,
-        titleEng: 'Project management 25',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'wordpress'],
-        iconPath: 'media/svg/term-cards/term-icon-2.svg',
-        keyword: 'wordpress'
-    },
-    {
-        id: 26,
-        titleEng: 'Project management 26',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'frontend'],
-        iconPath: 'media/svg/term-cards/term-icon-3.svg',
-        keyword: 'frontend'
-    },
-    {
-        id: 27,
-        titleEng: 'Project management 27',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'python'],
-        iconPath: 'media/svg/term-cards/term-icon-1.svg',
-        keyword: 'python'
-    },
-    {
-        id: 28,
-        titleEng: 'Project management 28',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'wordpress'],
-        iconPath: 'media/svg/term-cards/term-icon-2.svg',
-        keyword: 'wordpress'
-    },
-    {
-        id: 29,
-        titleEng: 'Project management 29',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'frontend'],
-        iconPath: 'media/svg/term-cards/term-icon-3.svg',
-        keyword: 'frontend'
-    },
-    {
-        id: 30,
-        titleEng: 'Project management 30',
-        titleGeo: 'პროექტის მენეჯმენტი',
-        Description: 'ადამიანური, მატერიალური და დროითი რესურსების დაგეგმვა',
-        hashTags: ['გრაფიკულიდიზაინი', 'frontend'],
-        iconPath: 'media/svg/term-cards/term-icon-3.svg',
-        keyword: 'frontend'
-    },
-    {
-        id: 31,
-        titleEng: 'Project',
-        titleGeo: 'პროექტი',
-        Description: 'პროექტი არის დროებითი საქმიანობა უნიკალური პროდუქტის, მომსახურების ან შედეგის შექმნის მიზნით',
-        hashTags: ['ციფრულიკომუნიკაციები', 'პროექტისმენეჯმენტი'],
-        iconPath: 'media/svg/term-cards/term-icon-3.svg',
-        keyword: 'frontend',
-        weekTerm: true
-    },
-]
-const alphabetArrGeo = ['ა', 'ბ', 'გ', 'დ', 'ე', 'ვ', 'ზ', 'თ', 'ი', 'კ', 'ლ', 'მ', 'ნ', 'ო', 'პ', 'ჟ', 'რ', 'ს', 'ტ', 'უ', 'ფ', 'ქ', 'ღ', 'ყ', 'შ', 'ჩ', 'ც', 'ძ', 'წ', 'ჭ', 'ხ', 'ჯ', 'ჰ']
-const alphabetArrEng = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z']
+import { data } from "./data.js";
+
+const cardsWrapper = document.querySelector(".term-cards-wrapper");
 let wordsCount = 7;
-const cardsWrapper = document.querySelector('.term-cards-wrapper')
-const members = [
-    {
-        memberId: 0,
-        memberName: 'სახელი',
-        memberLastname: 'გვარი',
-        memberPosition: 'Frontend Developer',
-        memberImagePath: 'media/png/person.png'
-    },
-    {
-        memberId: 1,
-        memberName: 'სახელი',
-        memberLastname:  'გვარი',
-        memberPosition: 'Backend Developer',
-        memberImagePath: 'media/png/person.png'
-    },
-    {
-        memberId: 2,
-        memberName: 'სახელი',
-        memberLastname: 'გვარი',
-        memberPosition: 'UI/UX Designer',
-        memberImagePath: 'media/png/person.png'
-    },
-    {
-        memberId: 3,
-        memberName: 'სახელი',
-        memberLastname:  'გვარი',
-        memberPosition: 'Project Manager',
-        memberImagePath: 'media/png/person.png'
-    },
-]
-if(document.location.pathname == '/unilab-dictionary/about.html') {
-    const teamWrapper = document.querySelector('.meet-team-wrapper')
-    const personName = document.querySelectorAll('.person-name')
-    const personLastname = document.querySelectorAll('.person-lastname')
-    const personLPosition = document.querySelectorAll('.person-position')
-  
-    renderMembers(members, teamWrapper)
 
-
-
-    function renderMembers (array, wrapper) {
-        wrapper.innerHTML = ''
-        array.forEach(el => {
-            wrapper.innerHTML += 
-            `
-                <div class="slick-slide-wrapper" style="background-image: url(${el.memberImagePath})"><div class="about-person"><span class="person-name">${el.memberName}</span><span class="person-lastname">${el.memberLastname}</span><span class="person-position">${el.memberPosition}</span></div></div>
-            `
-        })
-        // <img src=${el.memberImagePath} alt="team member">
-    }
-    $('.meet-team-wrapper').slick({
-        // autoplay: true,
-        infinite: true,
-        dots: false,
-        speed: 300,
-        slidesToShow: 4,
-        slidesToScroll: 2,
-        centerMode: true,
-        responsive: [
-    {
-      breakpoint: 1024,
-      settings: {
-        slidesToShow: 3,
-        slidesToScroll: 2,
-        infinite: true,
-        dots: true
-      }
-    },
-    {
-      breakpoint: 769,
-      settings: {
-        slidesToShow: 2,
-        slidesToScroll: 2,
-        infinite: true,
-        dots: true
-      }
-    },
-    {
-      breakpoint: 600,
-      settings: {
-        slidesToShow: 1,
-        slidesToScroll: 1,
-        dots: true
-      }
-    },
-    {
-      breakpoint: 480,
-      settings: {
-        slidesToShow: 1,
-        slidesToScroll: 1,
-        dots: true
-      }
-    }
-          // You can unslick at a given breakpoint now by adding:
-          // settings: "unslick"
-          // instead of a settings object
-        ]
-      });
-   
+const weekTermTitleEnd = document.querySelector(".term-title strong");
+const weekTermTitleGeo = document.querySelector(".term-title span");
+const weekTermDesc = document.querySelector(".week-term-description");
+const firstHashtag = document.querySelector(".hashtag-1");
+const secondHashtag = document.querySelector(".hashtag-2");
+const weekTerm = data.filter((term) => term.weekTerm);
+weekTermTitleEnd.textContent = weekTerm[0].titleEng;
+weekTermTitleGeo.textContent = weekTerm[0].titleGeo;
+weekTermDesc.textContent =
+  weekTerm[0].Description.split(" ").splice(0, 11).join(" ") + "...";
+firstHashtag.textContent = "#" + weekTerm[0].hashTags[0];
+secondHashtag.textContent = "#" + weekTerm[0].hashTags[1];
+console.log(weekTerm);
+const dataToRender = data.filter((item) => item.id < 3);
+renderData(dataToRender);
+if (document.body.offsetWidth < 769) {
+  const cards = document.querySelectorAll(".term-card");
+  cards.forEach((card) => {
+    card.classList.add("fullsize");
+  });
 }
-if (document.location.pathname == '/unilab-dictionary/dictionary-detailed.html') {
-    const termExample = document.querySelector('#termExample')
-    const exampleImg = document.querySelector('.example-img')
-    const dataToRender = data.filter(item => item.id < 6)
-    renderData(dataToRender)
-    if(exampleImg === null) {
-        termExample.style.maxWidth = '1000px';
-    }
-}
-if (document.location.pathname == '/unilab-dictionary/index.html') {
-    const weekTermTitleEnd = document.querySelector('.term-title strong')
-    const weekTermTitleGeo = document.querySelector('.term-title span')
-    const weekTermDesc = document.querySelector('.week-term-description')
-    const firstHashtag = document.querySelector('.hashtag-1')
-    const secondHashtag = document.querySelector('.hashtag-2')
-    weekTerm = data.filter(term => term.weekTerm)
-    weekTermTitleEnd.textContent = weekTerm[0].titleEng
-    weekTermTitleGeo.textContent = weekTerm[0].titleGeo
-    weekTermDesc.textContent = weekTerm[0].Description.split(' ').splice(0, 11).join(' ') + '...'
-    firstHashtag.textContent = '#' + weekTerm[0].hashTags[0]
-    secondHashtag.textContent = '#' + weekTerm[0].hashTags[1]
-    console.log(weekTerm)
-    const dataToRender = data.filter(item => item.id < 3)
-    renderData(dataToRender)
-    if (document.body.offsetWidth < 769) {
-        const cards = document.querySelectorAll('.term-card')
-        cards.forEach(card => {
-            card.classList.add('fullsize')
-        })
-    }
-}
-if (document.location.pathname == '/unilab-dictionary/dictionary.html') {
-    const swiperWrapper = document.querySelector('.swiper-wrapper');
-    const alphabetWrapper = document.querySelector('.alphabet-wrapper')
-    const alphabetSwitcher = document.querySelector('.alphabet-switcher')
-    const searchFilter = document.querySelector('#filter')
-    const resultCounter = document.querySelector('.search-result-counter')
-    const messageWrapper = document.querySelector('.counter-wrapper')
-    const notFoundMessage = document.querySelector('.not-found-message-wrapper')
-    const paginationWrapper = document.querySelector('.pagination')
-    let termsPerPage = 9
-    let paginationResult
-    if(document.body.offsetWidth < 480) {
-        termsPerPage = 3
-    }else if (document.body.offsetWidth < 991) {
-        termsPerPage = 6
 
-    }else {
-        termsPerPage = 9
+console.log(document.location.pathname);
 
-    }
-    const swiper = new Swiper('.swiper', {
-        // Navigation arrows
-        navigation: {
-          nextEl: '.swiper-button-next',
-          prevEl: '.swiper-button-prev',
-        },
-        // And if we need scrollbar
-        scrollbar: {
-          el: '.swiper-scrollbar',
-        },
-        slidesPerView: 10,
-        slidesPerGroup: 3
-      });
-    // initiate first page on window load
-    window.onload = () => {
-        // document.querySelector('.page-number').click()
-    }
-    // initiate first page on window load
-    // generating 9 cards per page
-    paginationRender(data)
-    function paginationRender (arr) {
-        paginationWrapper.innerHTML = ''
-        const roundedTermsNum = Math.ceil(arr.length / termsPerPage)
-        const doubleLeft = document.createElement('span')
-        doubleLeft.setAttribute('class', 'pagination-double-left-arrow')
-        doubleLeft.innerHTML = 
-        `
-            <<
-        `
-        paginationWrapper.append(doubleLeft)
-        const leftArrow = document.createElement('span')
-        leftArrow.setAttribute('class', 'pagination-left-arrow')
-        leftArrow.innerHTML = 
-        `
-        <svg xmlns="http://www.w3.org/2000/svg" width="12.51" height="8.893" viewBox="0 0 12.51 8.893">
-            <g id="noun_Arrow_2884045" transform="translate(12.51) rotate(90)">
-                <path id="Path_3" data-name="Path 3" d="M8.307,12.51.207,6.66a.5.5,0,0,1,0-.81L8.307,0l.586.811L1.354,6.255,8.893,11.7Z" transform="translate(0 0)" fill="#8a8fa3" opacity="0.63"/>
-            </g>
-        </svg>
-        `
-        paginationWrapper.append(leftArrow)
-        if(roundedTermsNum >= 1) {
-            for (let i = 1; i <= roundedTermsNum; i++) {
-                const pageNumber = document.createElement('span')
-                pageNumber.setAttribute('class', 'page-number')
-                pageNumber.innerText = i
-                paginationWrapper.append(pageNumber)
-                pageNumber.addEventListener('click', function () {
-                    document.querySelectorAll('.page-number').forEach(el => el.classList.remove('active-page'))
-                    this.classList.add('active-page')
-                    const pageNum = +this.innerText
-                    const startPoint = (pageNum - 1) * termsPerPage
-                    const endPoint = startPoint + termsPerPage
-                    paginationResult = arr.slice(startPoint, endPoint)
-                    cardsWrapper.innerHTML = ''
-                    renderData(paginationResult)
-                })
-            }
-        document.querySelector('.page-number').click()
-        }
-        const rightArrow = document.createElement('span')
-        rightArrow.setAttribute('class', 'pagination-right-arrow')
-        rightArrow.innerHTML = `
-        <svg xmlns="http://www.w3.org/2000/svg" width="12.51" height="8.893" viewBox="0 0 12.51 8.893">
-            <g id="noun_Arrow_2884045" transform="translate(0 8.893) rotate(-90)">
-                <path id="Path_3" data-name="Path 3" d="M8.307,0,.207,5.85a.5.5,0,0,0,0,.81l8.1,5.85.586-.811L1.354,6.255,8.893.81Z" fill="#8a8fa3" opacity="0.63"/>
-            </g>
-        </svg>
-        `
-        paginationWrapper.append(rightArrow)
-        leftArrow.addEventListener('click', () => {
-            console.log('left clicked')
-        })
-        rightArrow.addEventListener('click', () => {
-            console.log('right clicked')
-        })
-        const doubleRight = document.createElement('span')
-        doubleRight.setAttribute('class', 'pagination-double-right-arrow')
-        doubleRight.innerHTML = 
-        `
-           >>
-        `
-        paginationWrapper.append(doubleRight)
-    }
-    // generating 9 cards per page
-    // generate alphabet for desktop
-    alphabetGenerator(alphabetArrGeo)
-    // generate alphabet for desktop
-    // initiate slider
-    swiperInit(alphabetArrGeo)
-    // initiate slider
-    // adding active class to first letter
-    alphabetWrapper.childNodes[1].classList.add('active-letter')
-    swiperWrapper.childNodes[0].classList.add('active-letter')
-    // adding active class to first letter
-    lettersOnClick('.letter-box')
-    const switchFace = document.querySelector('.switch-face')
-    const background = document.querySelector('.alphabet-bg')
-    const switchTextEng = 'ENG'
-    const switchTextGeo = 'ქარ'
-    switchFace.innerText = switchTextEng
-    background.addEventListener('click', (e)=> {
-        e.stopImmediatePropagation()
-        alphabetSwitcher.click()
-    })
-    alphabetSwitcher.addEventListener('click', (e) => {
-        e.stopImmediatePropagation()
-        if (switchFace.innerText == switchTextEng) {
-            switchFace.innerText = switchTextGeo
-            alphabetWrapper.innerText = ''
-            alphabetGenerator(alphabetArrEng)
-            console.log(document.body.offsetWidth)
-            if(document.body.offsetWidth < 570) {
-                alphabetSwitcher.style.transform = 'translateX(25px)'
-            }else {
-                alphabetSwitcher.style.transform = 'translateX(34px)'
-            }
-            lettersOnClick('.letter-box')
-            swiperInit(alphabetArrEng)
-            alphabetWrapper.childNodes[0].classList.add('active-letter')
-            swiperWrapper.childNodes[0].classList.add('active-letter')
-        }
-        else {
-            switchFace.innerText = switchTextEng
-            alphabetWrapper.innerText = ''
-            alphabetGenerator(alphabetArrGeo)
-            alphabetSwitcher.style.transform = 'translateX(0px)'
-            lettersOnClick('.letter-box')
-            swiperInit(alphabetArrGeo)
-            alphabetWrapper.childNodes[0].classList.add('active-letter')
-            swiperWrapper.childNodes[0].classList.add('active-letter')
-        }
-
-    })
-    searchFilter.addEventListener('change', () => {
-        const searchFilterData = data.filter(item => item.keyword == searchFilter.value)
-        searchCounter(searchFilterData)
-        if (searchFilter.value !== 'default') {
-            cardsWrapper.innerHTML = ''
-            paginationRender(searchFilterData)
-        }
-    })
-
-    const search = document.querySelector('#dictionary-search')
-    search.addEventListener('input', (e) => {
-        cardsWrapper.style.justifyContent = 'space-around'
-        cardsWrapper.innerHTML = ''
-        const filteredData = data.filter(item => item.titleEng.toLowerCase().includes(e.target.value.toLowerCase()) || item.titleGeo.includes(e.target.value))
-        if(filteredData.length < 10 && filteredData.length !== 0) {
-            notFoundMessage.style.display = 'none'
-            searchCounter(filteredData)
-            paginationRender(filteredData)
-            if(filteredData.length === 1) {
-                cardsWrapper.style.justifyContent = 'flex-start'
-            }
-            paginationWrapper.style.display = 'block'
-        }else if (filteredData.length == 0){
-            notFoundMessage.style.display = 'flex'
-            searchCounter(filteredData)
-            paginationWrapper.style.display = 'none'
-        }
-        if (e.target.value == '') {
-            paginationRender(filteredData)
-            notFoundMessage.style.display = 'none'
-            paginationWrapper.style.display = 'block'
-            messageWrapper.style.display = 'none'
-        }else if (e.target.value !== '' && filteredData.length >= 10){
-            searchCounter(filteredData)
-            paginationRender(filteredData)
-            notFoundMessage.style.display = 'none'
-
-        }
-    })
-    function searchCounter(arr) {
-        if (arr != null) {
-            messageWrapper.style.display = 'block'
-            resultCounter.innerText = arr.length
-        }
-    }
-    function alphabetGenerator(alphabetArray) {
-        if (alphabetArray[0] == alphabetArrGeo[0]) {
-            alphabetArray.forEach((letter) => {
-                const letterSpan = document.createElement('span')
-                letterSpan.setAttribute('class', 'letter-box geo')
-                letterSpan.innerText = letter
-                alphabetWrapper.append(letterSpan)
-                letterSpan.addEventListener('click', (e) => {
-                    console.log(e.target);
-                })
-            })
-        } else {
-            alphabetArray.forEach((letter) => {
-                const letterSpan = document.createElement('span')
-                letterSpan.setAttribute('class', 'letter-box')
-                letterSpan.innerText = letter.toUpperCase()
-                alphabetWrapper.append(letterSpan)
-                letterSpan.addEventListener('click', (e) => {
-                    console.log(e.target);
-                })
-            })
-        }
-    }
-    function lettersOnClick(someClass) {
-        const letterBoxes = document.querySelectorAll(someClass)
-        letterBoxes.forEach(letterBox => {
-            letterBox.addEventListener('click', function () {
-                letterBoxes.forEach(letter => letter.classList.remove('active-letter'))
-                letterBox.classList.add('active-letter')
-            })
-        })
-    }
-    function swiperInit (arr) {
-        swiperWrapper.innerHTML = '';
-        arr.forEach(el => {
-            const upperCase = el.toUpperCase()
-            const letter = document.createElement('div')
-            letter.setAttribute('class', 'swiper-slide')
-            letter.innerText = upperCase
-            swiperWrapper.append(letter)
-            lettersOnClick('.swiper-slide')
-        })
-    }
-
-}
+// import { renderData } from "./dictionary.js";
 function renderData(array) {
-    if (document.body.offsetWidth < 769) {
-            wordsCount = 6
-        
-        }
-    if (document.body.offsetWidth < 480) {
-            wordsCount = 5
-        }
-    array.forEach((element) => {
-    cardsWrapper.innerHTML += ''
-        // console.log(element.Description)
-        const card = document.createElement('div')
-        card.setAttribute('class', 'term-card')
-        cardsWrapper.append(card)
-        card.innerHTML = `
-                  <div class="card-header">
-                      <span class="term-icon">
-                          <img src="${element.iconPath}" alt="third icon">
-                      </span>
-                      <h3 class="term-header-title"><span>${element.titleEng} -</span><span>${element.titleGeo}</span></h3>
-                  </div>
-                  <div class="card-body">
-                      <p class="term-description">${element.Description.split(' ').splice(0, wordsCount).join(' ') + "..."}</p>
-                  </div>
-                  <div class="card-footer">
-                      <div class="hashtag-keywords">
-                        <a href="dictionary.html"><span>#${element.hashTags[1]}</span></a>
-                        <a href="dictionary.html"><span>#${element.hashTags[0]}</span></a>
-                      </div>
-                      <div class="button-wrapper">
-                          <a href="#" class="see-details">ნახე სრულად</a>
-                      </div>
-                  </div>
-    `
-    })
+  if (document.body.offsetWidth < 769) {
+    wordsCount = 6;
+  }
+  if (document.body.offsetWidth < 480) {
+    wordsCount = 5;
+  }
+  array.forEach((element) => {
+    cardsWrapper.innerHTML += "";
+    // console.log(element.Description)
+    const card = document.createElement("div");
+    card.setAttribute("class", "term-card");
+    cardsWrapper.append(card);
+    card.innerHTML = `
+
+        <div class="card-header">
+            <span class="term-icon">
+                <img src="${element.iconPath}" ../alt="third icon">
+            </span>
+            <h3 class="term-header-title"><span>${element.titleEng} -</span><span>${element.titleGeo}</span></h3>
+        </div>
+
+        <div class="card-body">
+            <p class="term-description">${element.Description.split(" ").splice(0, wordsCount).join(" ") + "..."}</p>
+        </div>
+
+        <div class="card-footer">
+            <div class="hashtag-keywords">
+            <a href="dictionary.html"><span>#${element.hashTags[1]}</span></a>
+            <a href="dictionary.html"><span>#${element.hashTags[0]}</span></a>
+        </div>
+
+        <div class="button-wrapper">
+            <a href="#" class="see-details">ნახე სრულად</a>
+        </div>
+    `;
+  });
 }
+
+
+
+
+const search = document.querySelector('#main-search')
+search.addEventListener('input', (e) => {
+    const filteredData = data.filter(item => item.titleEng.toLowerCase().includes(e.target.value.toLowerCase()) || item.titleGeo.includes(e.target.value))
+  console.log(filteredData);
+
+
+
+
+
+
+
+
+
+
+
+
+    // if(filteredData.length < 10 && filteredData.length !== 0) {
+    //     notFoundMessage.style.display = 'none'
+    //     searchCounter(filteredData)
+    //     paginationRender(filteredData)
+    //     if(filteredData.length === 1) {
+    //         cardsWrapper.style.justifyContent = 'flex-start'
+    //     }
+    //     paginationWrapper.style.display = 'block'
+    // }else if (filteredData.length == 0){
+    //     notFoundMessage.style.display = 'flex'
+    //     searchCounter(filteredData)
+    //     paginationWrapper.style.display = 'none'
+    // }
+    // if (e.target.value == '') {
+    //     paginationRender(filteredData)
+    //     notFoundMessage.style.display = 'none'
+    //     paginationWrapper.style.display = 'block'
+    //     messageWrapper.style.display = 'none'
+    // }else if (e.target.value !== '' && filteredData.length >= 10){
+    //     searchCounter(filteredData)
+    //     paginationRender(filteredData)
+    //     notFoundMessage.style.display = 'none'
+
+    // }
+});
